requirements.txt
streamlit
openai
ollama
psycopg2-binary
pandas
matplotlib
seaborn